+++ native/GC/CHANGELOG.md
@@ -1,4 +1,5 @@
-## 1.0.3
+## 1.0.4
 
 ### In Android & iOS
-    - Added a public API to show or hide the language switch option.
+    - Added a public API (hideEndChatPopupWindow) to hide or show the end chat popup.
+    - Added a public API (hideLanguageMismatchPopupWindow) to hide or show the language mismatch popup.